# docs/swagger.yaml
components:
  schemas:
    model.Credentials:
      properties:
        email:
          form: email
          type: string
        password:
          form: password
          type: string
      required:
        - email
        - password
      type: object
    model.User:
      properties:
        email:
          form: email
          type: string
        name:
          form: name
          type: string
        password:
          form: password
          type: string
        picture_url:
          form: picture_url
          type: string
      required:
        - email
        - name
        - password
      type: object
    model.UserIdentified:
      properties:
        email:
          form: email
          type: string
        id:
          form: id
          type: integer
        name:
          form: name
          type: string
        password:
          form: password
          type: string
        picture_url:
          form: picture_url
          type: string
      required:
        - email
        - name
        - password
      type: object
    service.AuthResult:
      properties:
        email:
          form: email
          type: string
        id:
          form: id
          type: integer
        jwt:
          format: jwt
          type: string
        name:
          form: name
          type: string
        password:
          form: password
          type: string
        picture_url:
          form: picture_url
          type: string
      required:
        - email
        - name
        - password
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact:
    email: suhartanto@kekkon.nexus
    name: Ryan Suhartanto
    url: https://github.com/ryansuhartanto/koda-b8-backend2
  description: Go exercise implementing a simple REST server.
  license:
    name: MIT
  title: GoREST
  version: "1.0"
openapi: 3.1.0
paths:
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                - $ref: "#/components/schemas/model.Credentials"
                  description: Credentials
                  summary: body
          application/x-www-form-urlencoded:
            schema:
              type: string
          multipart/form-data:
            schema:
              type: object
        description: Credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/service.AuthResult"
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: invalid request body
        "401":
          content:
            application/json:
              schema:
                type: string
          description: email not registered
        "422":
          content:
            application/json:
              schema:
                type: string
          description: incorrect password
        "500":
          content:
            application/json:
              schema:
                type: string
          description: internal error
      summary: Log in
      tags:
        - auth
  /auth/register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                - $ref: "#/components/schemas/model.User"
                  description: New user
                  summary: body
          application/x-www-form-urlencoded:
            schema:
              type: string
          multipart/form-data:
            schema:
              type: object
        description: New user
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/service.AuthResult"
          description: Created
        "400":
          content:
            application/json:
              schema:
                type: string
          description: invalid request body
        "409":
          content:
            application/json:
              schema:
                type: string
          description: email already exists
        "500":
          content:
            application/json:
              schema:
                type: string
          description: internal error
      summary: Register a new user
      tags:
        - auth
  /users/:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/model.UserIdentified"
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                type: string
          description: internal error
      security:
        - BearerAuth: []
      summary: List users
      tags:
        - users
  /users/{id}:
    delete:
      parameters:
        - description: User ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: invalid request
        "500":
          content:
            application/json:
              schema:
                type: string
          description: internal error
      security:
        - BearerAuth: []
      summary: Delete a user
      tags:
        - users
    patch:
      parameters:
        - description: User ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                - $ref: "#/components/schemas/model.User"
                  description: User fields
                  summary: body
          application/x-www-form-urlencoded:
            schema:
              type: string
          multipart/form-data:
            schema:
              type: object
        description: User fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.UserIdentified"
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: invalid request
        "500":
          content:
            application/json:
              schema:
                type: string
          description: internal error
      security:
        - BearerAuth: []
      summary: Update a user
      tags:
        - users
  /users/{id}/picture:
    put:
      parameters:
        - description: User ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/octet-stream:
            schema:
              format: binary
              type: string
          application/x-www-form-urlencoded:
            schema:
              title: picture
              type: file
          image/*:
            schema:
              type: string
          multipart/form-data:
            schema:
              type: object
        description: Picture file (omit or send empty body to clear)
      responses:
        "200":
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: invalid request
        "413":
          content:
            application/json:
              schema:
                type: string
          description: file too large
        "422":
          content:
            application/json:
              schema:
                type: string
          description: unsupported image format
        "500":
          content:
            application/json:
              schema:
                type: string
          description: internal error
      security:
        - BearerAuth: []
      summary: Upload or clear a user's profile picture
      tags:
        - users
servers:
  - url: http://localhost:8080